+++ CHANGES.md
@@ -1,3 +1,15 @@
+1.0.3 / 2020-10-22
+==================
+
+  * [added] Helper function directus() can now be overritten via ServiceProvider
+
+1.0.2 / 2020-10-21
+==================
+
+  * [fixed] Helpers file location
+  * [fixed] Static instances
+  * [fixed] Use default connection with null
+
 1.0.0 / 2020-10-20
 ==================